minor image upload fixes, prevent image reuploads, and faster edit article action

+++ lib/articles/editArticle.action.ts
@@ -1,5 +1,5 @@
 "use server";
-import { ArticleFormState } from "@/data-types/Article";
+import { Article, ArticleFormState } from "@/data-types/Article";
 import db from "@/db";
 import { articleTable, contentTable } from "@/db/schema";
 import { eq } from "drizzle-orm";
@@ -10,6 +10,28 @@ import {
   removeImages,
 } from "./server-utils";
 
+const removeUnneededImages = async (
+  oldArticle: Article,
+  article: ArticleFormState
+): Promise<void | { error: string }> => {
+  const oldSrc: string[] = [
+    oldArticle.thumbnail,
+    ...oldArticle.contents
+      .map((c) => (c.type === "image" ? c.src || "" : ""))
+      .filter((s) => s !== ""),
+  ];
+
+  const newSrc = new Set<string>([
+    article.thumbnail.src || "",
+    ...article.contents
+      .map((c) => (c.type === "image" ? c.src || "" : ""))
+      .filter((s) => s !== ""),
+  ]);
+
+  const imagesToRemove = oldSrc.filter((x) => !newSrc.has(x));
+  await removeImages(imagesToRemove);
+};
+
 export const editArticle = async (
   articleId: number,
   article: ArticleFormState
@@ -18,26 +40,22 @@ export const editArticle = async (
   if (sessionError) return sessionError;
   console.log("starting editing article" + article.title);
 
+  if (!article.thumbnail.src) {
+    return { error: "Thumbnail is required" };
+  }
   try {
     const res1 = await getArticle(articleId, false);
     if ("error" in res1) return res1;
 
-    await db
-      .update(articleTable)
-      .set({ title: article.title, thumbnail: article.thumbnail.src })
-      .where(eq(articleTable.id, articleId));
-
-    await db.delete(contentTable).where(eq(contentTable.articleId, articleId));
+    await Promise.all([
+      db
+        .update(articleTable)
+        .set({ title: article.title, thumbnail: article.thumbnail.src })
+        .where(eq(articleTable.id, articleId)),
+      removeUnneededImages(res1.article, article),
+      db.delete(contentTable).where(eq(contentTable.articleId, articleId)),
+    ]);
     await insertContents(articleId, article);
-
-    const oldArticle = res1.article;
-    const oldSrcs = [
-      oldArticle.thumbnail,
-      ...article.contents
-        .map((c) => (c.type === "image" ? c.src : ""))
-        .filter((s) => s !== ""),
-    ];
-    await removeImages(oldSrcs);
   } catch (e) {
     if (e instanceof Error) {
       return { error: "Failed to edit article: " + e.message };
